split the redis file into one that can test that the tool is working and another that willl extract data
